requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
asyncpg>=0.29.0

# Vector database - 使用最新版本，兼容Mac/Windows/Linux
faiss-cpu>=1.8.0
numpy>=1.24.0,<2.0.0

# OpenAI compatible
openai>=1.3.7
tiktoken>=0.5.2

# Utilities
pyyaml>=6.0.1
aiohttp>=3.9.1
httpx>=0.25.2
tenacity>=8.2.3

# Monitoring
prometheus-client>=0.19.0

# Logging
loguru>=0.7.2

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
